misc/HtmlTestSuite/results/test-0248.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Forced Color Adjust: getComputedStyle for forced-color-adjust</title>
<link href="https://www.w3.org/TR/css-color-adjust-1/#forced-color-adjust-prop" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/computed-testcommon.js"/>
</head><body><div id="target"/>
<script>
  test_computed_value("forced-color-adjust", "auto");
  test_computed_value("forced-color-adjust", "none");
  test_computed_value("forced-color-adjust", "preserve-parent-color");
</script>
</body></html>
